Change variable declarations for MSVC and size issues
A slight change allows MSVC to compile the code as C. Additionally,
change the size of some variables to silence the casting warnings.

index and nread have been changed to size_t as they're array
offsets. In practice to_read is always going to be positive and won't
ever need 64 bits in a 32-bit machine.