Fix issues around multi-line headers
Always discard leading whitespace in a header value, even if it
is folded.

Pay attention to values of interesting headers (Connection,
Content-Length, etc.) even when they come on a continuation line.

Add a test case to check that requests and responses using only LF
to separate lines are handled correctly.