makefile: fix DESTDIR usage
DESTDIR is not supposed to be set inside other variables. It is
standard to have this variable in install/uninstall rules, so it
can be reliably set separately no matter what other variables are set
to.
This also avoids potential bugs with setting SONAME or seds on
installed files (like pkgconfig) which then might include the
temporary DESTDIR directory.

DESTDIR is really just for installing into a temporary directory or
a chroot, mostly used by package managers.