makefile: add CFLAGS to linking command
Although we compile the objects explicitly there are some CFLAGS
that may also affect linking, which is not always obvious.

This can also be a problem for toolchains that support multiple ABIs
and need to set CFLAGS=<abi selector>, which will cause linking
to either fail or produce an unusable executable/library.