add settings flags, refactor lenient_http_headers setting
A previous change during a security fix added the ability
to set a `lenient_http_headers` flag on the HTTP Parser
instance. The way it was done was a bit of a hack at the
time in order to avoid an ABI breaking change. This PR
adds a new `uint32_t` flags field to `http_parser_settings`
in order to provide a more robust and correct way of passing
these kinds of settings into a parser instance.