Fixed DESTDIR
Modified the Makefile so that it actually works. As it is now, symlinks
will point to the fully qualified path of wherever your DESTDIR was: so
you'd end up with symlinks like /usr/local/lib/libhttp_parser.so.2.7
pointing to
/tmp/http-parser-git/pkg/http-parser-git/usr/local/lib/libhttp_parser.so.2.7.1.
This caused me quite a bit of trouble when I was trying to package it,
but the current edits seem to work just fine after testing it in a few
locations.