Add an ability to disable body parsing
Currently the on_headers_complete callback can return 0 to proceed
with the response body, 1 to skip the body, and 2 to skip the body
and stop looking for further messages, and this decision on what to do
with the body can only be made inside the on_headers_complete callback.

This commit adds an ability to alter the parser behaviour later
after on_headers_complete pauses the parser and returns 0.