url: treat empty port as default
When parsing URLs, treat an empty port (eg `http://hostname:/`) as if it
were unspecified.  RFC 3986 says:

> URI producers and normalizers SHOULD omit the port component and its
> ":" delimiter if port is empty or if its value would be the same as
> that of the scheme's default.

(Emphasis on the "SHOULD" is mine.)  This indicates that URIs MAY be
produced with an empty port and the `:` delimiter.

Thus, we stop failing if we end host parsing at the port delimiter.